+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update project to build `net47` and `net48`.
 - Update `Images` with light and dark theme.
 - Update `ricaun.Revit.UI` with the image auto theme based in the `light` and `dark` in the file name.
+- Update `Exception` when null.
 ### Console
 - Throw exception when `RevitTestUtils` is empty.
 ### TestAdapter