+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.3.5] / 2024-07-22
+## [1.3.5] / 2024-07-22 - 2024-08-28
 ### Features
 - Update `Application` images to support `light` and `dark` theme.
 ### Application